oritQuew
OritQuew is not a standard or commonly used term in programming or data structures. It seems to be a typo or a placeholder name. If you are referring to a queue, it could be a misspelling of "queue" or "original queue". If you meant to create a queue data structure, you can implement it in Python as follows: